Refactor: 여러 구현체가 있을 수 있으므로 인터페이스를 통해 구현함
