+++ src/main/java/com/growup/pms/auth/service/mail/MailClient.java
@@ -1,0 +1,7 @@
+package com.growup.pms.auth.service.mail;
+
+import com.growup.pms.auth.service.dto.EmailDetails;
+
+public interface MailClient {
+    void sendEmail(EmailDetails emailDetails);
+}
